refactor: Structural output list network profiles
