+++ test_runner/src/main/kotlin/ftl/domain/ListNetworkProfiles.kt
@@ -1,13 +1,11 @@
 package ftl.domain
 
-import flank.common.logLn
 import ftl.api.fetchNetworkProfiles
-import ftl.environment.common.toCliTable
+import ftl.presentation.Output
 
-interface ListNetworkProfiles
+interface ListNetworkProfiles : Output
 
 operator fun ListNetworkProfiles.invoke() {
-    // TODO move toCliTable() and printing presentation layer during refactor of presentation after #1728
-    logLn("fetching available network profiles...")
-    logLn(fetchNetworkProfiles().toCliTable())
+    "fetching available network profiles...".out()
+    fetchNetworkProfiles().out()
 }
