+++ test_runner/src/main/kotlin/ftl/presentation/cli/firebase/test/environment/TestEnvironmentInfo.kt
@@ -8,6 +8,7 @@ import ftl.presentation.cli.firebase.test.android.orientations.toCliTable
 import ftl.presentation.cli.firebase.test.ipblocks.toCliTable
 import ftl.presentation.cli.firebase.test.locale.toCliTable
 import ftl.presentation.cli.firebase.test.providedsoftware.toCliTable
+import ftl.presentation.cli.firebase.test.networkprofiles.toCliTable
 
 fun TestEnvironment.Android.prepareOutputString() = buildString {
     appendLine(osVersions.toCliTable())
